Update to storage_key!
